+++ audio/whisper_functions.py
